backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.32.3
python-dotenv==1.0.0
gradio_client==1.13.3
huggingface_hub>=0.19.3
Pillow==10.4.0
